requirements.txt
pandas==1.4.4
py-ecg-detectors=1.3.2
matplotlib==3.5.1
numpy==1.22.3